Fixed warnings

+++ ArduinoAddons/Arduino_1.x.x/Melzi/cores/arduino/pins_arduino.c
@@ -385,27 +385,27 @@ const uint8_t PROGMEM digital_pin_to_timer_PGM[] = {
 const uint16_t PROGMEM port_to_mode_PGM[] =
 {
         NOT_A_PORT,
-    &DDRA,
-        &DDRB,
-        &DDRC,
-        &DDRD,
+		(uint16_t)&DDRA,
+        (uint16_t)&DDRB,
+        (uint16_t)&DDRC,
+        (uint16_t)&DDRD,
 };
 
 const uint16_t PROGMEM port_to_output_PGM[] =
 {
         NOT_A_PORT,
-        &PORTA,
-        &PORTB,
-        &PORTC,
-        &PORTD,
+        (uint16_t)&PORTA,
+        (uint16_t)&PORTB,
+        (uint16_t)&PORTC,
+        (uint16_t)&PORTD,
 };
 const uint16_t PROGMEM port_to_input_PGM[] =
 {
         NOT_A_PORT,
-        &PINA,
-        &PINB,
-        &PINC,
-        &PIND,
+        (uint16_t)&PINA,
+        (uint16_t)&PINB,
+        (uint16_t)&PINC,
+        (uint16_t)&PIND,
 };
 const uint8_t PROGMEM digital_pin_to_port_PGM[] =
 {
